# Optimize the performance of function bitmapAnd 