I hereby agree to the terms of the CLA available at: https://yandex.ru/legal/cla/?lang=en

Changelog category (leave one):
- Performance Improvement


Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Speed up bitmapAnd function

I found that when I use the function ```bitmapAnd(bt1, bt2)```, if the bt1 is type ```RoaringBitmap``` and bt2 has a small number of values which are saved in the ```SmallSet```, the speed is slower than the ```bitmapAnd(bt2, bt1)```

Comparing the performance of a simple query:
```
with (select bt1 from bitmap_local where event_id='2') as big_bitmap select bitmapCardinality(bitmapAnd(big_bitmap, bt1)) from bitmap_local where event_id>'2';

with (select bt1 from bitmap_local where event_id='2') as big_bitmap select bitmapCardinality(bitmapAnd(bt1, big_bitmap)) from bitmap_local where event_id>'2';
```

- the first query Elapsed: 0.037 sec. Processed 1.00 thousand rows, 540.13 KB (27.16 thousand rows/s., 14.64 MB/s.) 
- the second query Elapsed: 1.878 sec. Processed 1.00 thousand rows, 540.13 KB (533.42 rows/s., 287.54 KB/s.)

So I try to exchange the position of two args in this case. With the same query the final results are similar, both are about 0.035 sec.

IMPROVEMENT: **60x as fast**