@vdimir 2021-08-30
Thank you for you interest in ClickHouse!

Could you write performance test or provide full reproducible example with `CREATE` ?

I tested it on trivial case and it don't show difference between `bitmapAnd(a, b)`, and `bitmapAnd(b, a)`. 


<details>
<summary>create table</summary>

```
DROP TABLE IF EXISTS t1;
DROP TABLE IF EXISTS t2;

CREATE TABLE t1 (`b` AggregateFunction(groupBitmap, UInt64)) ENGINE = TinyLog;
CREATE TABLE t2 (`b` AggregateFunction(groupBitmap, UInt64)) ENGINE = TinyLog;


INSERT INTO t1 
WITH
    (SELECT groupArray(number) from numbers(30)) as s
SELECT bitmapBuild(s) as b;

INSERT INTO t2
WITH
    (SELECT groupArray(number) from numbers(100000000)) as s
SELECT bitmapBuild(s) as b;
```


</details>

```
$ clickhouse benchmark --query "SELECT bitmapAnd(t1.b, t2.b) FROM t1, t2"                  
...

Queries executed: 101.

localhost:9000, queries 4, QPS: 3.865, RPS: 7.730, MiB/s: 0.030, result RPS: 3.865, result MiB/s: 0.015.

0.000%		0.227 sec.	
10.000%		0.227 sec.	
20.000%		0.235 sec.	
30.000%		0.235 sec.	
40.000%		0.235 sec.	
50.000%		0.247 sec.	
60.000%		0.247 sec.	
70.000%		0.247 sec.	
80.000%		0.247 sec.	
90.000%		0.326 sec.	
95.000%		0.326 sec.	
99.000%		0.326 sec.	
99.900%		0.326 sec.	
99.990%		0.326 sec.	
```

```
$ clickhouse benchmark --query "SELECT bitmapAnd(t2.b, t1.b) FROM t1, t2"                

...

Queries executed: 102.

localhost:9000, queries 4, QPS: 3.740, RPS: 7.480, MiB/s: 0.029, result RPS: 3.740, result MiB/s: 0.015.

0.000%		0.147 sec.	
10.000%		0.147 sec.	
20.000%		0.159 sec.	
30.000%		0.159 sec.	
40.000%		0.159 sec.	
50.000%		0.207 sec.	
60.000%		0.207 sec.	
70.000%		0.207 sec.	
80.000%		0.207 sec.	
90.000%		0.557 sec.	
95.000%		0.557 sec.	
99.000%		0.557 sec.	
99.900%		0.557 sec.	
99.990%		0.557 sec.	
```