@UnamedRus 2021-08-30
Difference visible if one of bitmaps - constant.

```
WITH
    bitmapBuild(range(100000)) AS large,
    bitmapBuild(range(toUInt32(5))) AS small
SELECT count()
FROM numbers_mt(1000000)
WHERE NOT ignore(bitmapAnd(large, materialize(small)))

Query id: e5e2c8dc-def6-457f-a7fc-e02368b68db5

┌─count()─┐
│ 1000000 │
└─────────┘

1 rows in set. Elapsed: 0.963 sec. Processed 1.04 million rows, 8.32 MB (1.08 million rows/s., 8.64 MB/s.)

WITH
    bitmapBuild(range(100000)) AS large,
    bitmapBuild(range(toUInt32(5))) AS small
SELECT count()
FROM numbers_mt(1000000)
WHERE NOT ignore(bitmapAnd(materialize(small), large))

Query id: b9bc081b-0794-48f7-9d80-571e7b9bf3df

┌─count()─┐
│ 1000000 │
└─────────┘

1 rows in set. Elapsed: 0.181 sec. Processed 1.04 million rows, 8.32 MB (5.75 million rows/s., 46.02 MB/s.)
```