@OodounaikingoO 2021-08-31
I'm so glad to receive your email.

And here are my test statements:

for create table

create table arr_local (event_id String, arr_1 Array(UInt64), arr_2 Array(UInt64)) Engine = MergeTree() order by event_id;

create materialized view bitmap_local engine=MergeTree() order by event_id as select event_id, bitmapBuild(arr_1) as bt1, bitmapBuild(arr_2) as bt2 from arr_local;

INSERT INTO bitmap_local SELECT
    '1',
    bitmapBuild(n_array),
    bitmapBuild(n_array)
FROM
(
    SELECT groupArray(number) AS n_array
    FROM
    (
        SELECT number
        FROM system.numbers
        LIMIT 1
    )
)


for insert

INSERT INTO bitmap_local
select *
from
(
WITH
    (
    SELECT bitmapBuild(n_array)
    FROM
    (
        SELECT groupArray(number) AS n_array
        FROM
        (
            SELECT number
            FROM system.numbers
            LIMIT 90000000
        )
    )
    ) AS bitmap_1
SELECT '2', bitmap_1, bt1 FROM bitmap_local WHERE event_id = '1'
);




insert into bitmap_local
select *
from
(
with
    (
    select bt2
    from bitmap_local
    where event_id='2'
    ) as small_bitmap
select toString(number) as event_id, small_bitmap as bt1, small_bitmap as bt2 from system.numbers where number>2 limit 1000
);


for query

 with (select bt1 from bitmap_local where event_id='2') as big_bitmap select bitmapCardinality(bitmapAnd(big_bitmap, bt1)) from bitmap_local where event_id>'2';

 with (select bt1 from bitmap_local where event_id='2') as big_bitmap select bitmapCardinality(bitmapAnd(bt1, big_bitmap)) from bitmap_local where event_id>'2';


Because the test performance of a single line cannot be reflected, I tried multiple lines for the and operation.





________________________________
发件人: Vladimir C ***@***.***>
发送时间: 2021年8月30日 15:12
收件人: ClickHouse/ClickHouse ***@***.***>
抄送: dddounaiking ***@***.***>; Author ***@***.***>
主题: Re: [ClickHouse/ClickHouse] Optimize the performance of function bitmapAnd (#28332)


Thank you for you interest in ClickHouse!

Could you write performance test or provide full reproducible example with CREATE ?

I tested it on trivial case and it don't show difference between bitmapAnd(a, b), and bitmapAnd(b, a).

create table

DROP TABLE IF EXISTS t1;
DROP TABLE IF EXISTS t2;

CREATE TABLE t1 (`b` AggregateFunction(groupBitmap, UInt64)) ENGINE = TinyLog;
CREATE TABLE t2 (`b` AggregateFunction(groupBitmap, UInt64)) ENGINE = TinyLog;


INSERT INTO t1
WITH
    (SELECT groupArray(number) from numbers(30)) as s
SELECT bitmapBuild(s) as b;

INSERT INTO t2
WITH
    (SELECT groupArray(number) from numbers(100000000)) as s
SELECT bitmapBuild(s) as b;


$ clickhouse benchmark --query "SELECT bitmapAnd(t1.b, t2.b) FROM t1, t2"
...

Queries executed: 101.

localhost:9000, queries 4, QPS: 3.865, RPS: 7.730, MiB/s: 0.030, result RPS: 3.865, result MiB/s: 0.015.

0.000%          0.227 sec.
10.000%         0.227 sec.
20.000%         0.235 sec.
30.000%         0.235 sec.
40.000%         0.235 sec.
50.000%         0.247 sec.
60.000%         0.247 sec.
70.000%         0.247 sec.
80.000%         0.247 sec.
90.000%         0.326 sec.
95.000%         0.326 sec.
99.000%         0.326 sec.
99.900%         0.326 sec.
99.990%         0.326 sec.


$ clickhouse benchmark --query "SELECT bitmapAnd(t2.b, t1.b) FROM t1, t2"

...

Queries executed: 102.

localhost:9000, queries 4, QPS: 3.740, RPS: 7.480, MiB/s: 0.029, result RPS: 3.740, result MiB/s: 0.015.

0.000%          0.147 sec.
10.000%         0.147 sec.
20.000%         0.159 sec.
30.000%         0.159 sec.
40.000%         0.159 sec.
50.000%         0.207 sec.
60.000%         0.207 sec.
70.000%         0.207 sec.
80.000%         0.207 sec.
90.000%         0.557 sec.
95.000%         0.557 sec.
99.000%         0.557 sec.
99.900%         0.557 sec.
99.990%         0.557 sec.


―
You are receiving this because you authored the thread.
Reply to this email directly, view it on GitHub<https://github.com/ClickHouse/ClickHouse/pull/28332#issuecomment-908330888>, or unsubscribe<https://github.com/notifications/unsubscribe-auth/ANMXQASTEEHJDW3B3DVJG63T7N7SZANCNFSM5DBMYWPA>.
Triage notifications on the go with GitHub Mobile for iOS<https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675> or Android<https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub>.